move function to services
