+++ src/services/index.ts
@@ -8,6 +8,8 @@ import {
   subscribeToTokenListFactory,
   getTokensFactory,
   getTokenFromErc20Factory,
+  TokenFromExchange,
+  TokenFromErc20,
 } from './factories'
 
 import { logDebug } from 'utils'
@@ -35,11 +37,11 @@ export const subscribeToTokenList = subscribeToTokenListFactory(apis)
 
 export const getTokens = getTokensFactory(apis)
 
-export interface AddTokenToListParams {
+export interface TokenAndNetwork {
   networkId: number
   tokenAddress: string
 }
-export interface AddTokenToExchangeParams extends AddTokenToListParams {
+export interface AddTokenToExchangeParams extends TokenAndNetwork {
   userAddress: string
 }
 
@@ -60,7 +62,7 @@ type AddTokenResult = AddTokenResultSuccess | AddTokenResultFailure
 
 export const isTokenAddedSuccess = (result: AddTokenResult): result is AddTokenResultSuccess => result.success
 
-export const addTokenToList = async ({ networkId, tokenAddress }: AddTokenToListParams): Promise<AddTokenResult> => {
+export const addTokenToList = async ({ networkId, tokenAddress }: TokenAndNetwork): Promise<AddTokenResult> => {
   const checkSumAddress = web3.utils.toChecksumAddress(tokenAddress)
   const { token } = await getTokenFromExchangeByAddress({ networkId, tokenAddress: checkSumAddress })
   if (token) {
@@ -114,5 +116,55 @@ export const addTokenToExchange = async ({
     tokenList: tokenListApi.getTokens(networkId),
     token,
     error,
+  }
+}
+
+type ValidResons =
+  | TokenFromExchange.NOT_ERC20
+  | TokenFromExchange.NOT_REGISTERED_ON_CONTRACT
+  | TokenFromExchange.NOT_IN_TOKEN_LIST
+
+export interface FetchTokenResult {
+  token: TokenDetails | TokenFromErc20 | null
+  reason: ValidResons | null
+}
+
+export const fetchTokenData = async (params: TokenAndNetwork): Promise<FetchTokenResult> => {
+  try {
+    const [tokenInExchange, erc20Token] = await Promise.all([
+      // check if registered token
+      exchangeApi.hasToken(params),
+      // get ERC20 data
+      getTokenFromErc20(params),
+    ])
+
+    if (!tokenInExchange)
+      return {
+        token: erc20Token,
+        reason: erc20Token ? TokenFromExchange.NOT_REGISTERED_ON_CONTRACT : TokenFromExchange.NOT_ERC20,
+      }
+
+    if (!erc20Token)
+      return {
+        token: null,
+        reason: TokenFromExchange.NOT_ERC20,
+      }
+
+    // get registered id
+    const tokenId = await exchangeApi.getTokenIdByAddress(params)
+
+    return {
+      token: {
+        ...erc20Token,
+        id: tokenId,
+      },
+      reason: TokenFromExchange.NOT_IN_TOKEN_LIST,
+    }
+  } catch (error) {
+    logDebug('Error fetching token', params, error)
+    return {
+      token: null,
+      reason: null,
+    }
   }
 }
